requirements.txt
gradio>=4.0.0
openai>=1.0.0
tiktoken>=0.3.0
PyPDF2>=3.0.0
pydub>=0.25.1
requests>=2.28.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0